refactor: recategorize miscategorized tests and remove redundant tests
Moves:
- boundaries → contracts: conversation-state-results, tool-factory-shapes
- contracts → boundaries: execute-tool-boundary
- contracts → behavior: consume-stream-completion, stop-conditions
- composition → behavior: input-normalization, format-compatibility
- composition → integration: next-turn-params-flow, orchestrator-executor
- integration → behavior: conversation-state-format, stop-conditions-step-result
- pipelines → behavior: async-resolution-pipeline, orchestrator-utility-chain
- pipelines → integration: format-round-trip

Splits:
- pipelines/claude-conversion-deep → dispatch (routing test), behavior (annotations), integration (unsupported content)
- composition/stream-data-pipeline → contracts/tool-call-response-consistency (kept contract test, removed redundant stream tests)

Removals:
- dispatch/execute-tool-dispatch (redundant with behavior/tool-execution)
- integration/reusable-stream-consumers test 1 (redundant with behavior/reusable-stream)
