+++ tests/behavior/async-resolution-pipeline.test.ts
