+++ tests/behavior/consume-stream-completion.test.ts
