+++ tests/behavior/conversation-state-format.test.ts
