+++ tests/behavior/input-normalization.test.ts
