+++ tests/behavior/stop-conditions-step-result.test.ts
