+++ tests/behavior/stop-conditions.test.ts
