+++ tests/contracts/conversation-state-results.test.ts
