+++ tests/contracts/tool-factory-shapes.test.ts
